UserMenu to load its own model instead of passing it via app
